Extract CLI formatting helpers before command routing split
This is the first low-risk CLI decomposition step. Rich output formatting and
spinner/JSON helpers now live under `mcp_video.cli`, while `mcp_video.__main__`
still owns parser construction and command dispatch. Command names, flags,
lazy imports, and text/JSON output behavior are unchanged.

Constraint: Preserve `mcp_video.__main__:main` and all existing CLI command behavior
Rejected: Split parser and handlers in the same change | surface tests are still landing and a smaller step is easier to review
Confidence: high
Scope-risk: narrow
Tested: python3 -m pytest tests/test_cli.py tests/test_doctor.py -q --tb=short; python3 -m pytest tests/test_cli.py::TestCLIVersion::test_version_flag tests/test_cli.py::TestCLIInfo::test_info_outputs_json tests/test_doctor.py -q --tb=short; python3 -m mcp_video --version; python3 -m mcp_video --help; python3 -m mcp_video doctor --json; python3 -m ruff check mcp_video/__main__.py mcp_video/cli tests/test_cli.py; python3 -m ruff format --check mcp_video/__main__.py mcp_video/cli
Not-tested: full non-slow suite after CLI helper extraction
