+++ mcp_video/cli/__init__.py
@@ -1,0 +1,1 @@
+"""CLI helpers for mcp-video."""
